Move backports to separate webpack config
