+++ webpack/externals.js
@@ -1,0 +1,11 @@
+module.exports = {
+	// This is necessary for Gutenberg to work.
+	tinymce: "window.tinymce",
+
+	yoastseo: "window.yoast.analysis",
+	"yoast-components": "window.yoast.components",
+	react: "React",
+	"react-dom": "ReactDOM",
+
+	lodash: "window.lodash",
+};
